Persist scores when quiz is finished, load scores in high scores view, all using CoreData

+++ Gone Phishin'/Gone Phishin'/Classes/GPQuiz.h
@@ -23,7 +23,7 @@
 - (NSString *)getImageNameForQuestion:(int)n;
 - (int)numQuestions;
 - (int)numCorrect;
-- (float)percentCorrect;
+- (double)percentCorrect;
 
 - (void)setAnswerForQuestion:(int)n :(BOOL)trusts;
